+++ kubejs/startup_scripts/gtceu/Blocks and Items/crafting_componets.js
@@ -1,0 +1,8 @@
+/*GTCEuStartupEvents.craftingComponents(event => {
+//Plates
+    event.modifyUnificationEntry(CraftingComponent.PLATE, GTValues.UEV, new UnificationEntry('plate', 'gtceu:stabilized_oganesson_plate'))
+    event.modifyUnificationEntry(CraftingComponent.PLATE, GTValues.UIV, new UnificationEntry('plate', 'gtceu:shadow_radiance_plate'))
+
+//Cables
+    event.modifyUnificationEntry(CraftingComponent.CABLE, GTValues.UEV, new UnificationEntry('plate', 'gtceu:end_steel_single_cable'))
+})*/